Shuffle segments around
