+++ fints/segments/journal.py
@@ -3,7 +3,7 @@ from fints.formals import (
     ReferenceMessage, Response
 )
 
-from . import FinTS3Segment, ParameterSegment, ParameterSegment_22
+from ._base import FinTS3Segment, ParameterSegment, ParameterSegment_22
 
 class HKPRO3(FinTS3Segment):
     """Statusprotokoll anfordern, version 3
